requirements.txt
# =============================================================================
# Ares Docker Agent - Python Dependencies
# =============================================================================
# Last security audit: December 2024
# All versions pinned to address Docker Scout CVEs

# Web Framework (CVE-2024-47874, CVE-2024-24762, CVE-2024-53981, CVE-2025-62727 fixes)
fastapi==0.124.0
uvicorn[standard]==0.34.3
starlette==0.49.1
python-multipart==0.0.20
jinja2==3.1.6

# Security (CVE-2024-26130 fix)
bcrypt==4.2.1
cryptography==44.0.1
pynacl==1.5.0  # For Curve25519 (WireGuard keys)

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# gRPC (for platform registration) - CVE-2025-4565 fix
grpcio==1.68.1
grpcio-tools==1.68.1
protobuf==5.29.5

# HTTP Client (with HTTP/2 support for TLS 1.3)
httpx[http2]==0.28.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Session management
itsdangerous==2.2.0
